requirements.txt
wget==3.2
pyyaml==6.0.2
openai==1.82.1
tiktoken==0.9.0
markdown==3.8
beautifulsoup4==4.13.4
gitpython==3.1.44